// backend/src/logs/enums/log-type.enum.ts
export enum LogType {
    USER_CREATED = 'USER_CREATED',
    USER_LOGIN = 'USER_LOGIN',
    TRANSACTION_CREATED = 'TRANSACTION_CREATED',
    TRANSACTION_PROCESSING = 'TRANSACTION_PROCESSING',
    TRANSACTION_COMPLETED = 'TRANSACTION_COMPLETED',
    TRANSACTION_FAILED = 'TRANSACTION_FAILED',
    RABBITMQ_ERROR = 'RABBITMQ_ERROR',
    RABBITMQ_SUCCESS = 'RABBITMQ_SUCCESS',
    ERROR = 'ERROR',
}